server/requirements.txt
django
psycopg2-binary
gunicorn
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
coverage